tested final changes worked on the celery worker and startup/kill_scripts

+++ VideoPlayer.py
@@ -9,6 +9,7 @@ class VideoPlayer:
         self.path_to_video = None
         self.video_is_playing = False
         self.player = None
+        self.video_is_stopped = True
 
     def set_video(self, path):
         self.path_to_video = path
@@ -18,8 +19,11 @@ class VideoPlayer:
     def play_video(self):
         self.player = Popen(['omxplayer', '-o', 'hdmi', '-b', self.path_to_video], shell=False)
         self.video_is_playing = True
+        self.video_is_stopped = False
 
     def stop_video(self):
-        os.system('killall omxplayer.bin')
-        self.video_is_playing = False
-        self.player = None
+        if not self.video_is_stopped:
+            os.system('killall omxplayer.bin')
+            self.video_is_playing = False
+            self.video_is_stopped = True
+            self.player = None
